Convert shortnames to FQDN during ingestion (#4708)
* Convert shortnames to FQDN during ingestion


* missing dot

* undo


* go nits
